Modularize the code in Ticket.tsx and remove the issuesFeed
