+++ src/components/Constants.tsx
@@ -1,9 +1,4 @@
 //@ts-nocheck
-
-const prodUrl = import.meta.env.VITE_API_PROD_BASE_URL;
-
-export const BASE_URL = prodUrl || "http://localhost:5001";
-
 export const EDIT_SVG = (
 	<svg height="16px" viewBox="0 -960 960 960" width="16px" fill="#e8eaed">
 		<path d="M192.63-192.63h42.2l449.32-448.94-42.17-42.17-449.35 448.88v42.23Zm-47.89 47.89v-109.61l546.7-547.21q6.68-5.87 15.46-9.58 8.78-3.7 18.48-3.7 9.23 0 18.03 3.49 8.81 3.5 16.42 9.99l42.14 42.37q6.57 6.86 9.93 15.81 3.36 8.96 3.36 17.74 0 9.75-3.45 18.62-3.45 8.87-9.87 15.53L254.35-144.74H144.74Zm622.61-581.07-41.13-40.87 41.13 40.87ZM663.07-662.4l-21.09-21.34 42.17 42.17-21.08-20.83Z" />
@@ -53,10 +48,30 @@ export const DOWN_SVG = (
 
 import { StylesConfig } from "react-select";
 
-type OptionType = {
+const prodUrl = import.meta.env.VITE_API_PROD_BASE_URL;
+
+export const BASE_URL = prodUrl || "http://localhost:5001";
+
+export const statusOptions = [
+	{ value: "New", label: "New", icon: NEW_SVG },
+	{ value: "in-progress", label: "In Progress", icon: IN_PROGRESS_SVG },
+	{ value: "resolved", label: "Resolved", icon: RESOLVED_SVG },
+];
+
+export const formatOptionLabel = ({
+	label,
+	icon,
+}: {
 	label: string;
-	value: string;
-};
+	icon: React.ReactNode;
+}) => (
+	<div style={{ display: "flex", alignItems: "center" }}>
+		{icon}
+		<span style={{ marginLeft: 8 }}>{label}</span>
+	</div>
+);
+
+import { OptionType } from "../types/ticketTypes";
 
 export const customStyles: StylesConfig<OptionType, false> = {
 	control: (provided, state) => ({
