+++ src/components/CustomerView.tsx
@@ -1,6 +1,5 @@
 import React, { useState } from "react";
 import TicketSubmissionForm from "./TicketSubmissionForm";
-import IssuesFeed from "./IssuesFeed";
 
 const CustomerView: React.FC = () => {
 	const [submissionToggled, setSubmissionToggled] = useState<boolean>(false);
@@ -10,7 +9,6 @@ const CustomerView: React.FC = () => {
 				submissionToggled={submissionToggled}
 				setSubmissionToggled={setSubmissionToggled}
 			/>
-			<IssuesFeed refresh={submissionToggled} />
 		</>
 	);
 };
